requirements.txt
# Core
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3
pytz==2024.2

# Data
duckdb==1.1.3
pandas==2.2.3
pyarrow==18.0.0
openpyxl==3.1.5

# Sources
yfinance==0.2.51
websockets==13.1
fredapi==0.5.2
pyshp==3.0.3

# AI
google-genai>=1.0

# Scheduler
apscheduler==3.10.4

# Testing
pytest==8.3.4
pytest-mock==3.14.0
responses==0.25.3
